requirements.txt
openai
redis
requests
dotenv
python-telegram-bot